+++ src/components/index.js
@@ -1,0 +1,4 @@
+import MenuItems from './MenuItems/MenuItems';
+import RestaurantDetail from './RestaurantDetail/RestaurantDetail';
+
+export { MenuItems, RestaurantDetail };
