+++ src/containers/CategoriesContainer/CategoriesContainer.test.jsx
