+++ src/containers/RegionsContainer/RegionsContainer.jsx
@@ -1,11 +1,8 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import {
-  selectRegion,
-  loadRestaurants,
-} from './actions';
+import { loadRestaurants, selectRegion } from '../../redux/actions';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
