+++ src/containers/RestaurantContainer/RestaurantContainer.jsx
@@ -2,13 +2,11 @@ import { useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import RestaurantDetail from './RestaurantDetail';
+import { RestaurantDetail } from '../../components';
 
-import {
-  loadRestaurant,
-} from './actions';
+import { loadRestaurant } from '../../redux/actions';
 
-import { get } from './utils';
+import { get } from '../../utils/utils';
 
 export default function RestaurantContainer({ restaurantId }) {
   const dispatch = useDispatch();
