+++ src/containers/index.js
@@ -1,0 +1,11 @@
+import CategoriesContainer from './CategoriesContainer/CategoriesContainer';
+import RegionsContainer from './RegionsContainer/RegionsContainer';
+import RestaurantContainer from './RestaurantContainer/RestaurantContainer';
+import RestaurantsContainer from './RestuarantsContainer/RestaurantsContainer';
+
+export {
+  CategoriesContainer,
+  RegionsContainer,
+  RestaurantContainer,
+  RestaurantsContainer,
+};
