+++ src/index.jsx
@@ -8,7 +8,7 @@ import { Provider } from 'react-redux';
 
 import App from './App';
 
-import store from './store';
+import store from './redux/store';
 
 ReactDOM.render(
   (
