+++ src/pages/index.js
@@ -1,0 +1,7 @@
+import AboutPage from './AboutPage/AboutPage';
+import HomePage from './HomePage/HomePage';
+import NotFoundPage from './NotFoundPage/NotFoundPage';
+import RestaurantPage from './RestaurantPage/RestaurantPage';
+import RestaurantsPage from './RestaurantsPage/RestaurantsPage';
+
+export { HomePage, AboutPage, NotFoundPage, RestaurantPage, RestaurantsPage };
